Refactor: Move Redis URI parsing functions to config.py
- Move parse_redis_uri() and set_redis_env_from_config() from main.py to config.py
- These functions belong in config.py as they handle configuration logic
- Clean up main.py by removing configuration-related code
- Import functions from config.py in main.py
- Maintain all functionality while improving code organization
