+++ src/main.py
@@ -1,6 +1,5 @@
 import sys
 import os
-import urllib.parse
 import click
 
 from src.common.connection import RedisConnectionManager
@@ -16,7 +15,7 @@ import src.tools.sorted_set
 import src.tools.set
 import src.tools.stream
 import src.tools.pub_sub
-from src.common.config import MCP_TRANSPORT
+from src.common.config import MCP_TRANSPORT, parse_redis_uri, set_redis_env_from_config
 
 
 class RedisMCPServer:
@@ -25,89 +24,6 @@ class RedisMCPServer:
 
     def run(self):
         mcp.run(transport=MCP_TRANSPORT)
-
-def parse_redis_uri(uri: str) -> dict:
-    """Parse a Redis URI and return connection parameters."""
-    parsed = urllib.parse.urlparse(uri)
-
-    config = {}
-
-    # Scheme determines SSL
-    if parsed.scheme == 'rediss':
-        config['ssl'] = True
-    elif parsed.scheme == 'redis':
-        config['ssl'] = False
-    else:
-        raise ValueError(f"Unsupported scheme: {parsed.scheme}")
-
-    # Host and port
-    config['host'] = parsed.hostname or '127.0.0.1'
-    config['port'] = parsed.port or 6379
-
-    # Database
-    if parsed.path and parsed.path != '/':
-        try:
-            config['db'] = int(parsed.path.lstrip('/'))
-        except ValueError:
-            config['db'] = 0
-    else:
-        config['db'] = 0
-
-    # Authentication
-    if parsed.username:
-        config['username'] = parsed.username
-    if parsed.password:
-        config['password'] = parsed.password
-
-    # Parse query parameters for SSL and other options
-    if parsed.query:
-        query_params = urllib.parse.parse_qs(parsed.query)
-
-        # Handle SSL parameters
-        if 'ssl_cert_reqs' in query_params:
-            config['ssl_cert_reqs'] = query_params['ssl_cert_reqs'][0]
-        if 'ssl_ca_certs' in query_params:
-            config['ssl_ca_certs'] = query_params['ssl_ca_certs'][0]
-        if 'ssl_ca_path' in query_params:
-            config['ssl_ca_path'] = query_params['ssl_ca_path'][0]
-        if 'ssl_keyfile' in query_params:
-            config['ssl_keyfile'] = query_params['ssl_keyfile'][0]
-        if 'ssl_certfile' in query_params:
-            config['ssl_certfile'] = query_params['ssl_certfile'][0]
-
-        # Handle other parameters
-        if 'db' in query_params:
-            try:
-                config['db'] = int(query_params['db'][0])
-            except ValueError:
-                pass
-
-    return config
-
-
-def set_redis_env_from_config(config: dict):
-    """Set environment variables from Redis configuration."""
-    env_mapping = {
-        'host': 'REDIS_HOST',
-        'port': 'REDIS_PORT',
-        'db': 'REDIS_DB',
-        'username': 'REDIS_USERNAME',
-        'password': 'REDIS_PWD',
-        'ssl': 'REDIS_SSL',
-        'ssl_ca_path': 'REDIS_SSL_CA_PATH',
-        'ssl_keyfile': 'REDIS_SSL_KEYFILE',
-        'ssl_certfile': 'REDIS_SSL_CERTFILE',
-        'ssl_cert_reqs': 'REDIS_SSL_CERT_REQS',
-        'ssl_ca_certs': 'REDIS_SSL_CA_CERTS',
-        'cluster_mode': 'REDIS_CLUSTER_MODE'
-    }
-
-    for key, env_var in env_mapping.items():
-        if key in config:
-            value = config[key]
-            if isinstance(value, bool):
-                value = 'true' if value else 'false'
-            os.environ[env_var] = str(value)
 
 
 @click.command()
